feat: add plugin interface
