+++ .gitignore
@@ -1,3 +1,6 @@
+cmd/bootkube/bootkube
+*.so
+
 _output
 _pluton_temp
 vendor-v[0-9].[0-9].[0-9]/
