+++ cmd/render/plugin/default/asset/asset.go
