+++ cmd/render/plugin/default/asset/asset_test.go
