+++ cmd/render/plugin/default/asset/images.go
