+++ cmd/render/plugin/default/asset/internal/templates.go
