+++ pkg/recovery/recover.go
@@ -17,14 +17,13 @@ import (
 	"reflect"
 
 	"github.com/ghodss/yaml"
+	"github.com/kubernetes-sigs/bootkube/cmd/render/plugin/default/asset"
 	v1apps "k8s.io/api/apps/v1"
-	"k8s.io/api/core/v1"
+	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
-
-	"github.com/kubernetes-sigs/bootkube/pkg/asset"
 )
 
 const (
