+++ pkg/recovery/recover_test.go
@@ -4,12 +4,11 @@ import (
 	"reflect"
 	"testing"
 
+	"github.com/kubernetes-sigs/bootkube/cmd/render/plugin/default/asset"
 	v1apps "k8s.io/api/apps/v1"
-	"k8s.io/api/core/v1"
+	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-
-	"github.com/kubernetes-sigs/bootkube/pkg/asset"
 )
 
 var (
